requirements.txt
pandas==1.5.3
boto3==1.29.6
pytest==6.2.5
pytest-mock==3.7.0
PyYAML==6.0
coloredlogs==15.0.1